requirements.txt
Flask>=2.0.2
jinja2>=3.0.0
google-cloud-secret-manager>=2.8.0